+++ src/backend/resolvers/mutations/aquilaConfig/aquilaConfig/delete.js
@@ -1,0 +1,11 @@
+const { ObjectId } = require('mongodb')
+
+const deleteAquilaConfig = (
+  parent,
+  { input: { _id } },
+  { pulseCoreDb }
+) => pulseCoreDb.collection('businessObjects.aquilaConfigs')
+  .findOneAndDelete({ _id: ObjectId(_id) })
+  .then(({ value }) => value)
+
+module.exports = deleteAquilaConfig