+++ src/backend/resolvers/mutations/aquilaConfig/aquilaConfig/index.js
@@ -1,0 +1,9 @@
+const createAquilaConfig = require('./create')
+const updateAquilaConfig = require('./update')
+const deleteAquilaConfig = require('./delete')
+
+module.exports = {
+  createAquilaConfig,
+  updateAquilaConfig,
+  deleteAquilaConfig,
+}