+++ src/backend/resolvers/mutations/aquilaConfig/field/index.js
@@ -1,0 +1,9 @@
+const createAquilaConfigField = require('./create')
+const updateAquilaConfigField = require('./update')
+const deleteAquilaConfigField = require('./delete')
+
+module.exports = {
+  createAquilaConfigField,
+  updateAquilaConfigField,
+  deleteAquilaConfigField,
+}