+++ src/backend/resolvers/mutations/aquilaConfig/index.js
@@ -1,0 +1,7 @@
+const aquilaConfigResolvers = require('./aquilaConfig')
+const fieldResolvers = require('./field')
+
+module.exports = {
+  ...aquilaConfigResolvers,
+  ...fieldResolvers,
+}