+++ src/backend/resolvers/queries/aquila/index.js
@@ -1,0 +1,5 @@
+const aquilaConfigs = require('./aquilaConfigs')
+
+module.exports = {
+  aquilaConfigs,
+}