+++ src/backend/typeDefs/mutations/aquilaConfig/aquilaConfig/index.js
@@ -1,0 +1,9 @@
+const createAquilaConfigTypeDefs = require('./create')
+const deleteAquilaConfigTypeDefs = require('./delete')
+const updateAquilaConfigTypeDefs = require('./update')
+
+module.exports = [
+  createAquilaConfigTypeDefs,
+  deleteAquilaConfigTypeDefs,
+  updateAquilaConfigTypeDefs,
+]