+++ src/backend/typeDefs/mutations/aquilaConfig/field/delete.js
@@ -1,0 +1,10 @@
+const { gql } = require('apollo-server-express')
+
+const deleteAquilaConfigFieldTypeDefs = gql`
+  input DeleteAquilaConfigFieldInput {
+    _id: ID!
+    aquilaConfigId: ID!
+  }
+`
+
+module.exports = deleteAquilaConfigFieldTypeDefs